requirements.txt
dify_plugin
azure-storage-blob==12.24.0
